Move `CommitAuthor` to `operations.rs`
