+++ src/github/api/mod.rs
@@ -17,6 +17,8 @@ use client::GithubRepositoryClient;
 pub mod client;
 pub(crate) mod operations;
 
+pub use operations::CommitAuthor;
+
 /// GitHub requests will be timed out after this time by default
 pub const DEFAULT_REQUEST_TIMEOUT: Duration = Duration::from_secs(10);
